add set nav speed threshold cmd

+++ src/command/set_dgps_mode.rs
@@ -18,7 +18,7 @@ impl Command for SetDgpsMode {
     type Response = Ack;
 
     fn encode(&self) -> Result<PmtkPacket, PmtkError> {
-        let data_field = encode_data_field([self.mode as u32]);
+        let data_field = encode_data_field([self.mode as u8]);
         PmtkPacket::new_command(Self::PKT_TYPE, Some(data_field))
     }
 }
